pyproject.toml
[tool.poetry]
name = "dorylus"
version = "0.1.0"
description = ""
authors = ["samoslab <hanyh2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "1.9.0"
llama-index = "^0.10.15"
fastapi = "^0.102.0"
chromadb = "^0.4.9"
jupyter = "^1.0.0"
pymongo = "^4.5.0"
beanie = "^1.19.2"
fastapi-jwt = "^0.1.12"
fastapi-jwt-auth = "^0.5.0"
eth-abi = "^4.0.0"
pypdf2 = "^3.0.1"
web3 = "^6.9.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
py-solc-x = "^1.1.1"
base58 = "^2.1.1"
pyyaml = "^6.0.1"
python-socketio = "5.11.0"
websockets = "^11.0.3"
fastapi-sso = "^0.9.1"
python-ulid = "^2.2.0"
pyautogen = "^0.1.6"
pytest = "^7.4.2"
fastapi-mail = "^1.4.1"
email-validate = "^1.1.2"
python-decouple = "^3.8"
eth-account = "^0.9.0"
aes = "^1.2.0"
apscheduler = "^3.10.4"
cryptography = "^41.0.5"
redis = "^5.0.1"
pillow = "^10.1.0"
loguru = "^0.7.2"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.1"
llama-index-readers-file = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
